FileAPI/filelist/fileAPI_blob_slice.html

<!DOCTYPE html>
<html>
<head>
    <title>fileAPI blob test:</title>
    <link rel="author" href="mailto:jfengsky@gmail.com"/>
    <link rel="help" href="http://www.w3.org/TR/FileAPI/#dfn-file"/>
    <meta name="flags" content="dom"/>
    <meta name="assert" content="test new blob size and type. test blob slice. safari not support" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="stylesheet" href="/resources/testharness.css"/>
</head>
<body>
<p>fileAPI blob and blob slice test: test new blob size and type. test blob slice. safari not support</p>

<div id="log"></div>

<div id="J_demo"></div>
<script>
    var emptyBlob = new Blob();
    var content = ['blob', 'abcd'];
    var blobType = {"type": "text/xml"};
    var Blob2 = new Blob(content, blobType);




    test(function(){
        assert_equals(emptyBlob.size, 0, 'blob size is zero when create');
    }, 'blob size is zero when create');

    test(function(){
        assert_equals(emptyBlob.type, '', 'blob type is empyt when create');
    }, 'blob type is empyt when create');

    test(function(){
        assert_equals(Blob2.size, content.join('').length, 'blob size equal first argument length');
    }, 'blob size equal first argument length');

    var blobSliceMaxSize = Blob2.slice(10);
    test(function(){
        assert_equals(blobSliceMaxSize.size, 0, 'slice start more than source size return 0');
    },'slice start more than source size return 0');

    var blobSlicelessSize = Blob2.slice(0, 2);
    test(function(){
        assert_less_than(blobSlicelessSize.size, Blob2.size, 'slice start less than source size return the data between slice')
    }, 'slice start less than source size return the data between slice');


</script>
</body>
</html>
